fixed issue where form and questions appeared at the same time. form is gone after submission
